tests/wpt/tests/encoding-detection/ja-EUC-JP-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>ja EUC-JP late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "EUC-JP", 'Expected EUC-JP');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/ja-EUC-JP-late.sub.html");
</script>
